// lib/api/generated/models/models-paginated-gig-creator-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Huerray Backend
 * Comprehensive API for connecting content creators with brands for marketing campaigns and gig opportunities
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@thrypes.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ModelsGigCreatorResponse } from './models-gig-creator-response';
// May contain unused imports in some cases
// @ts-ignore
import type { ModelsPagination } from './models-pagination';

export interface ModelsPaginatedGigCreatorResponse {
    'data'?: Array<ModelsGigCreatorResponse>;
    'message'?: string;
    'pagination'?: ModelsPagination;
    'success'?: boolean;
}